Replace descriptor interface #15
